Move menu contents to Menu.js
